✨ 모바일 탐지 전역 파일 생성
