+++ src/components/common/Header.js
@@ -1,6 +1,8 @@
 import { React, useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { styled } from "styled-components";
+import useMobileDetection from "./mobileDetection";
+
 import profile from "../../assets/header/profile.png";
 import logo from "../../assets/header/logo.png";
 import mobilelogo from "../../assets/header/mobilelogo.png";
@@ -10,29 +12,9 @@ import HomeMenu from "./HomeMenu";
 
 const Header = () => {
   const navigate = useNavigate();
+  const isMobile = useMobileDetection();
 
-  const [isMobile, setisMobile] = useState(false);
   const [isHomeMenuOpen, setIsHomeMenuOpen] = useState(false);
-
-  // 리사이즈 이벤트를 감지하여 가로 길이에 따라 모바일 여부 결정
-  const resizingHandler = () => {
-    if (window.innerWidth <= 768) {
-      setisMobile(true);
-    } else {
-      setisMobile(false);
-    }
-  };
-
-  useEffect(() => {
-    if (window.innerWidth <= 1023) {
-      setisMobile(true);
-    }
-    window.addEventListener("resize", resizingHandler);
-    return () => {
-      // 메모리 누수를 줄이기 위한 removeEvent
-      window.removeEventListener("resize", resizingHandler);
-    };
-  }, []);
 
   const openModal = () => {
     setIsHomeMenuOpen(!isHomeMenuOpen);
